+++ final_project_frontend/src/components/RoomSelector.scss
@@ -1,0 +1,27 @@
+.community-selector {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 10px;
+  padding: 10px;
+  background-color: #f9f9f9;
+  border-radius: 10px;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
+
+  button {
+    background-color: #4A90E2;
+    color: white;
+    border: none;
+    padding: 10px 20px;
+    border-radius: 20px;
+    cursor: pointer;
+    transition: background-color 0.2s;
+
+    &.selected {
+      background-color: #357ABD;
+    }
+
+    &:hover {
+      background-color: #357ABD;
+    }
+  }
+}
